Copy content from 18-index.html to 20-index.html Explanation:
Duplicated the content of '18-index.html' into '20-index.html'.
Added navigation links in the <nav> tags:
Created a link to '/' with the text 'Home'.
Added anchors to 'services', 'works', 'about', 'latest_news', 'testimonials', and 'contact' with corresponding text for each section.
The anchor links are currently placeholders and will be made functional in the CSS project.