Copy content from 21-index.html to 22-index.html Explanation:
Duplicated the content of '21-index.html' into '22-index.html'.
Added links as follows:
In the Hero section, created a link with the text 'Get started' redirecting to '#'.
In the About Us section, created a link with the text 'Learn more about us' redirecting to 'about.html'.
In the Contact section, created a link with the text 'Get in touch' redirecting to 'contact.html'.