Copy content from 27-index.html to 29-index.html Explanation:
Duplicated the content of '27-index.html' into '29-index.html'.
Updated the Testimonials section:
Replaced the text in the first article with a blockquote containing 'I am completely blown away. Thanks to Techium, we've just launched our 5th website!' and cited the author as Yuri Y.
Replaced the text in the second article with a blockquote containing 'Thank you so much for your help. Techium company is awesome!' and cited the author as Dorrie S.
Replaced the text in the third article with a blockquote containing 'I love your system. Definitely worth the investment. I'd be lost without Techium company.' and cited the author as Sven H.