Copy content from 28-styleguide.html to 30-styleguide.html Explanation:
Duplicated the content of '28-styleguide.html' into '30-styleguide.html'.
Added examples of inline quote and blockquote:
Created a new section titled 'Blockquotes' with a level 2 heading.
Added an inline quote with the text 'Stay hungry. Stay foolish.'
Added a multiline blockquote with the text from Kanye West: 'I will be the leader of a company that ends up being worth billions of dollars, because I got the answers. I understand culture. I am the nucleus. I think that’s a responsibility that I have, to push possibilities, to show people, this is the level that things could be at.' and cited the author as Kanye West, Musician.