Copy content from 32-styleguide.html to 33-styleguide.html Explanation:
Duplicated the content of '32-styleguide.html' into '33-styleguide.html'.
Added a new section titled 'Table' with a level 2 heading.
Created an HTML table to reproduce the visual representation provided in the instructions.
Set the 'scope' attribute of the <th> tags for 'Title', 'Director', and 'Release Date' to 'col'.
Set the 'scope' attribute of the <th> tags for the movie names to 'row'.